+++ hcpworks/src/hcp_controller.ts
@@ -318,7 +318,12 @@ export class HCPController {
     const renderer = new SVGRenderer(svgContent.getName(), parseInfo4Render)
       .setModuleMeta(meta)
       .setSvgBgColor(this.configManager.getConfigSvgBgColor())
-      .setWireColorTable(this.configManager.getConfigWireColorTable());
+      .setWireColorTable(this.configManager.getConfigWireColorTable())
+      .setDisplayOptions(
+        this.configManager.getConfigShowName(),
+        this.configManager.getConfigShowScope(),
+        this.configManager.getConfigShowKind()
+      );
     const svgText = renderer.render();
 
     return svgContent.setSvgContent(svgText);